Fix: Emit RowChanged() when setting a table cell value.
This affects everyone using the uiprivTableModelSetCellValue()
Prior to this patch rows were not reliably updated when editing
cell values in a table.
Most notably: double clicking a checkbox in test/page16 on unix
would get the checkbox stuck in the wrong state (until gtk does
a full redraw for whatever reason).