Add table column IDs matching the column add order/index.
To be able to identify table columns individually use the order
that columns have been added to the table as the ID.
First added column: 0, second: 1, ...

It would be nice to have user defined column IDs in the future,
both cocoa and gtk implement these. This will incur an API
breakage though.